add strategy pattern
