+++ src/app/services/newsletter/newsletter.ts
@@ -1,0 +1,3 @@
+export interface INewsletter {
+    read(useCache: boolean): Promise<any>;
+}
