+++ src/main/routes/index.routes.ts
@@ -1,7 +1,9 @@
 import { Router } from 'express'
-import { FilipedeschampsController } from '../../app/presentation/controllers/filipedeschampsController'
+import { DeschampsNewsletter } from 'src/app/services/newsletter/deschampsNewsletter'
+import { NewsletterController } from '../../app/presentation/controllers/newsletterController'
 import { adaptRoute } from '../adapters/express.adapter'
 
 export default (router: Router): void => {
-  router.get('/news', adaptRoute(new FilipedeschampsController()))
+  router.get('/news', adaptRoute(new NewsletterController(new DeschampsNewsletter)))
+  router.get('/filipedeschamps', adaptRoute(new NewsletterController(new DeschampsNewsletter)))
 }
